test(ui): split spectrum renderer specs
Refs #2859
